plateforme-back/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.3.1
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.109.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
googleapis-common-protos==1.73.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.27.2
idna==3.11
itsdangerous==2.2.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
passlib==1.7.4
pillow==12.1.1
protobuf==5.29.6
proto-plus==1.27.1
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1_modules==0.4.2
pyasn1==0.6.3
pycparser==3.0
pydantic_core==2.41.5
pydantic==2.12.5
pyparsing==3.3.2
pypdf==6.9.1
python-docx==1.2.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
reportlab==4.4.10
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.35.1
tqdm==4.67.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.27.0
